gnhf 5: Audited the whole codebase with three parallel Opus reviewers and fixed what they found — a process-aborting broken-pipe bug, an unbounded reply, two contradictory refusal contracts, a 44% wasted JSON payload, whole-file reads used as existence probes — answered both open notebook Questions (one with a new archive cascade, rewritten after review found three defects in it), and cut the suite from 490 to 458 tests while covering ~20 more behaviours.

+++ crates/anb/src/cli.rs
@@ -251,7 +251,7 @@ pub struct AddArgs {
     pub draft: DraftArgs,
     /// 0–4, 0 the most urgent.
     #[arg(long)]
-    pub priority: Option<u8>,
+    pub priority: Option<u32>,
 }
 
 #[derive(Args)]
@@ -297,7 +297,7 @@ pub struct EditArgs {
     pub from: Option<String>,
     /// 0–4, 0 the most urgent.
     #[arg(long)]
-    pub priority: Option<u8>,
+    pub priority: Option<u32>,
     /// The explicit resurfacing date.
     #[arg(long)]
     pub review_by: Option<String>,
